+++ src/graphql_clj/parser.clj
@@ -2,154 +2,102 @@
   (:require [instaparse.core :as insta]
             [clojure.java.io :as io]
             [camel-snake-kebab.core :refer [->kebab-case]]
-            [clojure.string :as str]
             [clojure.set :as set]))
-
-(def whitespace (insta/parser "whitespace = #'\\s+'"))
 
 (def graphql-bnf "graphql.bnf")
 
-(def ^:private parse- (insta/parser (io/resource graphql-bnf)))
-
-(defn parse-debug
-  [stmt]
-  (insta/parse parse- stmt :partial true))
+(def ^:private parser-fn (insta/parser (io/resource graphql-bnf)))
 
 (defn- parse-statement
   "Parse graphql statement, hiccup format syntax tree will be returned for a valid graphql statement.
    An instance of instaparse.gll.Failure will be returned for parsing error."
   [stmt]
-  (parse- stmt))
+  (parser-fn stmt))
 
-(defn- transform-type-names [_ & args]
-  (let [names (into {} args)
-        type-names (:type-names names)
-        type-name (:name names)]
-    [:type-names (conj type-names type-name)]))
-
+(defn- to-object-value [_ & args] (->> (map #(vector (:name %) (:value %)) args) (into {})))
+(defn- to-default-value [_ arg] (set/rename-keys arg {:value :default-value}))
 (defn- args->map [_ & args] (into {} args))
 (defn- to-ident [_ v] v)
 (defn- to-map [k & args] {k (into {} args)})
 (defn- to-vec [k & args] [k (vec args)])
-(defn- to-val [k v] [k v])
-(defn- to-type-system-type [k & args] (into {:type-system-type (keyword (str/replace (name k) #"-definition" ""))} args))
-(defn- to-unwrapped-val [k v] [k (second v)])
-(defn- to-list [_ & args]
-  (let [{:keys [type-field-type]} (into {} args)]
-    {:kind :LIST :inner-type (set/rename-keys type-field-type {:name :type-name})}))
-
-(defn- add-required [_ arg]
-  (cond
-    (map? arg) (assoc arg :required true)
-    (and (vector? arg) (= :named-type (first arg))) {:type-name (last arg) :required true}
-    :else (throw (ex-info "unexpected to-required arg" arg))))
-
-(defn- to-singular-and-plural [k & args]
-  (let [base       (name k)
-        singular-k (keyword (subs base 0 (- (count base) 1)))
-        arg-map    (into {} args)]
-    [k (conj (k arg-map) (singular-k arg-map))]))
-
-(defn- to-document [_ & args]
-  (let [operation-definitions (filter #(= :operation-definition (:type %)) args) ; FIXME: assume there is only one operation-definition
-        type-definitions (filter #(= :type-system-definition (:type %)) args)
-        fragment-definitions (filter #(= :fragment-definition (:type %)) args)
-        fragments (reduce (fn reduce-fragments [v fragment]
-                            (let [name (:fragment-name fragment)]
-                              (assert name "fragment-name is NULL!")
-                              (assoc v name fragment)))
-                          {} fragment-definitions)]
-    {:operation-definitions   operation-definitions
-     :type-system-definitions type-definitions
-     :fragments               fragments}))
-
-(defn- to-operation-definition [_ & args]
-  (merge {:type           :operation-definition
-          :operation-type {:type "query"}}                  ; default operation as query
-         (into {} args)))
-
-(defn- to-name-value-pair [k [_ name] [_ value]]
-  (assert name "Name is NULL!")
-  [name value])
-
-(defn- to-unwrapped-name [k [_ name] & value-args]
-  (assert name "Name is NULL!")
-  [name (into {} value-args)])
-
-(defn- to-fragment-definition [_ & args]
-  (let [{:keys [fragment-name] :as definition} (into {} args)]
-    (assert fragment-name "fragment name is NULL for fragment-definition!")
-    (assoc definition :type :fragment-definition)))
-
-(def type-system-type->kind
-  {:type      :OBJECT
-   :input     :INPUT_OBJECT
-   :union     :UNION
-   :enum      :ENUM
-   :interface :INTERFACE
-   :directive :DIRECTIVE
-   :schema    :SCHEMA})
-
-(defn- to-type-system-definition [_ {:keys [type-system-type schema-types] :as definition}]
-  (-> (dissoc definition :schema-types)
-      (set/rename-keys {:name              :type-name
-                        :type-fields       :fields
-                        :enum-fields       :fields
-                        :input-type-fields :fields
-                        :directive-on-name :on})
-      (merge schema-types)
-      (assoc :type :type-system-definition
-             :kind (get type-system-type->kind type-system-type))))
-
+(defn- to-val [k v] {k v})
+(defn- to-type [_ v] (to-val :type v))
+(defn- to-enum-type [_ v] (to-val :enum-type v))
 (defn- parse-int    [_ v] (Integer/parseInt v))
 (defn- parse-double [_ v] (Double. v))
 (defn- parse-string [_ & args] (clojure.string/join (map second args)))
 (defn- parse-bool   [_ v] (= "true" v))
+(defn- unwrap-name [k v] {k (:name v)})
+(defn- to-one-or-more [_ & args] {:values (mapv :value args)})
+(defn- to-type-names [_ & args] {:type-names (mapv :type-name args)})
+(defn- to-list [_ arg] {:node-type :list :inner-type arg :kind :LIST})
+(defn- add-required [_ arg] (assoc arg :required true))
+(defn- to-document [_ & args] (group-by :section args))
 
-(defn- to-kv-map [k & args]
-  (let [{:keys [name type type-field-type] :as m} (set/rename-keys (into {} args) {:variable-name :name})]
-    (assert name (format "Name is NULL for (%s) (%s)!" k args))
-    [name (-> m
-              (dissoc :type-field-type :type :name)
-              (merge type-field-type type)
-              (set/rename-keys {:type-field-arguments        :arguments
-                                :type-field-argument-default :default-value
-                                :name                        :type-name
-                                :named-type                  :type-name}))]))
+(defn- to-type-system-type [k & args]
+  (-> (into {:node-type k} args)
+      (set/rename-keys {:type-field-arguments        :arguments
+                        :type-field-argument-default :default-value})))
+
+(defn- to-operation-definition [_ & args]
+  (merge {:section        :operation-definitions
+          :operation-type {:type "query"}}                  ; default operation as query
+         (into {} args)))
+
+(defn- to-fragment-definition [k & args]
+  (let [{:keys [name] :as definition} (into {} args)]
+    (assert name "fragment name is NULL for fragment-definition!")
+    (assoc definition :node-type k :section :fragment-definitions)))
+
+(def node-type->kind
+  {:type-definition      :OBJECT
+   :input-definition     :INPUT_OBJECT
+   :union-definition     :UNION
+   :enum-definition      :ENUM
+   :interface-definition :INTERFACE
+   :directive-definition :DIRECTIVE
+   :schema-definition    :SCHEMA})
+
+(defn- to-type-system-definition [_ definition]
+  (-> (assoc definition :section :type-system-definitions)
+      (set/rename-keys {:type-fields       :fields
+                        :enum-fields       :fields
+                        :input-type-fields :fields
+                        :directive-on-name :on})
+      (assoc :kind (get node-type->kind (:node-type definition)))))
 
 (def ^:private transformations
   "Map from transformation functions to tree tags.
    This map gets rendered into the format expected by instaparse, e.g.: {:TreeTag fn}
    Use :k to specify non-standard names for the first argument to the relevant transformation function"
-  {{:f to-ident}                  #{:Definition :SchemaType :DirectiveName :ArgumentValue :Type}
-   {:f to-document}               #{:Document}
-   {:f to-operation-definition}   #{:OperationDefinition}
-   {:f to-map}                    #{:OperationType :Selection :Field :Arguments :Directive :FragmentSpread :InlineFragment :SchemaTypes :QueryType :MutationType :DirectiveOnName :EnumField :TypeImplements :TypeFieldVariable :TypeFieldArguments :TypeFieldType :TypeFields :InputTypeFields :VariableDefinitions}
-   {:f to-val}                    #{:Name :Value :TypeCondition :EnumValue :TypeFieldVariableDefault :TypeFieldArgumentDefault}
-   {:f to-val :k :type}           #{:Query :Mutation}
-   {:f to-val :k :enum-type}      #{:EnumTypeInt}
-   {:f to-vec}                    #{:SelectionSet}
-   {:f to-name-value-pair}        #{:ObjectField}
-   {:f to-unwrapped-name}         #{:Argument}
-   {:f to-kv-map}                 #{:TypeField :InputTypeField :TypeFieldArgument :VariableDefinition}
-   {:f parse-int}                 #{:IntValue}
-   {:f parse-double}              #{:FloatValue}
-   {:f parse-string}              #{:StringValue}
-   {:f parse-bool}                #{:BooleanValue}
-   {:f to-fragment-definition}    #{:FragmentDefinition}
-   {:f to-unwrapped-val}          #{:NamedType :FragmentName :FragmentType :DefaultValue :Alias :VariableName}
-   {:f to-type-system-type}       #{:InterfaceDefinition :EnumDefinition :UnionDefinition :SchemaDefinition :InputDefinition :DirectiveDefinition :ScalarDefinition :TypeExtensionDefinition :TypeDefinition}
-   {:f to-type-system-definition} #{:TypeSystemDefinition}
-   {:f to-singular-and-plural}    #{:EnumFields :TypeFieldVariables}
-   {:f add-required}              #{:TypeFieldTypeRequired :NonNullType}
-   {:f transform-type-names}      #{:TypeNames :UnionTypeNames}
-   {:f to-list}                   #{:ListTypeName}
-   {:f args->map}                 #{:EnumType :ObjectValue}})
+  {to-ident                  #{:Definition :SchemaType :DirectiveName :ArgumentValue :ListValue :TypeFieldType :Type :Selection :EnumValue}
+   to-document               #{:Document}
+   to-operation-definition   #{:OperationDefinition}
+   to-map                    #{:OperationType :QueryType :MutationType :DirectiveOnName :Implements}
+   to-val                    #{:Name :Value :TypeCondition :TypeFieldArgumentDefault :ListType}
+   unwrap-name               #{:TypeName :ArgumentName :FieldName :VariableName}
+   to-type                   #{:Query :Mutation}
+   to-enum-type              #{:EnumTypeInt}
+   to-vec                    #{:SelectionSet :TypeFields :InputTypeFields :TypeFieldArguments :VariableDefinitions :Arguments :Directives :EnumFields}
+   parse-int                 #{:IntValue}
+   parse-double              #{:FloatValue}
+   parse-string              #{:StringValue}
+   parse-bool                #{:BooleanValue}
+   to-fragment-definition    #{:FragmentDefinition}
+   to-type-system-type       #{:InterfaceDefinition :EnumDefinition :UnionDefinition :SchemaDefinition :InputDefinition :DirectiveDefinition :ScalarDefinition :TypeExtensionDefinition :TypeDefinition :TypeField :InputTypeField :TypeFieldArgument :VariableDefinition :Argument :Directive :EnumField :Field :FragmentSpread :InlineFragment}
+   to-type-system-definition #{:TypeSystemDefinition}
+   add-required              #{:TypeFieldTypeRequired :NonNullType}
+   to-type-names             #{:TypeNames :UnionTypeNames}
+   to-one-or-more            #{:OneOrMoreValue}
+   to-list                   #{:ListTypeName}
+   args->map                 #{:EnumType :ObjectField :FragmentName :FragmentType :Alias :SchemaTypes}
+   to-object-value           #{:ObjectValue}
+   to-default-value          #{:DefaultValue}})
 
 (defn- render-transformation-fns
   "Invert the map of functions to tree tags (so instaparse receives tree tags to functions)."
-  [{:keys [f k]} v]
-  (map #(vector % (partial f (or k (->kebab-case %)))) v))
+  [f v]
+  (map #(vector % (partial f (->kebab-case %))) v))
 
 (def ^:private transformation-map
   "Map of tree tags to transformation functions that operate on the children."