+++ test/graphql_clj/introspection_test.clj
@@ -1,12 +1,11 @@
 (ns graphql-clj.introspection-test
   (:require [clojure.test :refer :all]
-            [clojure.java.io :as io]
             [graphql-clj.type :as type]
             [graphql-clj.parser :as parser]
-            [graphql-clj.resolver :as resolver]
-            [graphql-clj.executor :as executor]))
+            [graphql-clj.executor :as executor]
+            [graphql-clj.introspection :as intro]))
 
-(def simple-user-schema
+(def user-schema-str
   "type User {
   name: String
   nickname: String
@@ -22,19 +21,28 @@ schema {
   query: QueryRoot
 }")
 
-(defn- create-test-schema
-  [type-spec]
-  (-> type-spec
-      (parser/parse)
-      (type/create-schema)))
-
+(defn- create-test-schema [type-spec]
+  (type/create-schema (parser/parse type-spec) intro/introspection-schema))
 
 (deftest test-schema-introspection
-  (let [simple-schema (create-test-schema simple-user-schema)
-        introspection-schema (create-test-schema (slurp (io/resource "introspection.schema")))
-        schema (type/inject-introspection-schema simple-schema introspection-schema)
+  (let [schema (create-test-schema user-schema-str)
         resolver-fn nil
         context nil
         query "query { __schema { types {name kind} }}"
         result (executor/execute context schema resolver-fn query)]
-    (is (not (nil? (:data result))))))
+    (is (= (-> result :data (update-in ["__schema" "types"] set))
+           {"__schema" {"types" #{{"name" "QueryRoot" "kind" :OBJECT}
+                                  {"name" "User" "kind" :OBJECT}
+                                  {"name" "String" "kind" :SCALAR}
+                                  {"name" "Int" "kind" :SCALAR}
+                                  {"name" "Float" "kind" :SCALAR}
+                                  {"name" "Boolean" "kind" :SCALAR}
+                                  {"name" "ID" "kind" :SCALAR}
+                                  {"name" "__Schema" "kind" :OBJECT}
+                                  {"name" "__Type" "kind" :OBJECT}
+                                  {"name" "__TypeKind" "kind" :ENUM}
+                                  {"name" "__Field" "kind" :OBJECT}
+                                  {"name" "__InputValue" "kind" :OBJECT}
+                                  {"name" "__EnumValue" "kind" :OBJECT}
+                                  {"name" "__Directive" "kind" :OBJECT}
+                                  {"name" "__DirectiveLocation" "kind" :ENUM}}}}))))